server/requirement.txt
python-dotenv==1.0.1
python-multipart==0.0.9
simpel_captcha==0.1.4
APScheduler==3.10.4

fastapi==0.115.0
fastapi-cache2==0.2.2
pydantic-settings==2.5.2
tortoise-orm==0.21.6

aiosmtplib==3.0.2
aiofiles==24.1.0

asyncpg==0.29.0
aiomysql==0.2.0
aioredis==2.0.1

redis==5.0.8

oss2==2.19.0
qiniu==7.13.2
cos-python-sdk-v5==1.9.31

alibabacloud_dysmsapi20170525==3.0.0
tencentcloud-sdk-python==3.0.1233

python-weixin==0.5.0
wechatpayv3==1.2.52
python-alipay-sdk==3.3.0
